complete migration of ED settings to SUIVarCollection

+++ src/common/libcommon.cc
@@ -80,12 +80,12 @@ tokens( const string& s_, const char* sep)
 
 void
 agh::str::
-decompose_double( double value, float *mantissa, int *exponent)
+decompose_double( double value, double *mantissa, int *exponent)
 {
 	char buf[32];
 	snprintf( buf, 31, "%e", value);
 	*strchr( buf, 'e') = '|';
-	sscanf( buf, "%f|%d", mantissa, exponent);
+	sscanf( buf, "%lf|%d", mantissa, exponent);
 }
